requirements.txt
google-cloud-bigquery>=3.10.0
google-cloud-pubsub
